requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
flask==2.3.3
fonttools==4.42.1
gunicorn==21.2.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
threadpoolctl==3.2.0
werkzeug==2.3.7
zipp==3.16.2
